+++ packages/terra-hookshot/CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 ### Fixed
 * Mobile zoom on safari
 * Half px value causing blurriness
+* Disable Listener dependency on target
 
 Unreleased
 ----------